V2.7.2-alpha Fixed Bug Preventing Outputs Turning On During Night Hours
Previously none of the outputs could be turned on or off automatically during the night hours (22:00-07:00). This meant that if you set the light bar to turn on at 6AM, it would not turn on unless you pressed a button to turn the screen back on.

Now all of the outputs can be switched on and off regardless of the time of day or whether the screen is on or off.