V2.7.3-alpha Temporarily Fixed Bug Causing Weekday to Glitch
Previously the alarm would show the day of the week but when it changed from Saturday to Sunday the screen would glitch and the day of the week would not be displayed.

I found that this was because I was using the day() function instead of the weekday() function. The day function actually returns the day of the month and not the day of the week. As a result it was looking for values beyond the size of the daysOfTheWeek array.

To fix this I changed from using the day() function to using the weekday() function. However, this takes into consideration the month and year. Unfortunately I cannot get the month and year using the NTPClient library. 

To temporarily fix this I found a random month and year where the day of the week fitted with the correct day of the week. This is not a good fix but if you set it correctly and program the alarm it will display the correct day of the week.