V2.7.4-alpha Fixed Bug Causing Weekday to Glitch
I have switched to using a different NTPClient library (included). This library includes the getFormattedDate() function. Using this function I was able to extract the day, month and year which I could then use to set the local time. This ensures that the alarm clock always knows the correct date and therefore it will always display the correct day of the week with no glitches.