V2.0.9-release Fixed Issue #2: On/Off Time Logic
Previously the logic in `Device::itIsOnTime()` could not handle situations where the device needed to turn on in the evening, and off in the morning (like a phone charging). This was because "onTime > offTime" and therefore "onTime <= now < offTime" could never be true.

Now `Device::itIsOnTime()` has been updated to check whether "onTime > offTime", and if so, check if it is either after the on time at night, or before the off time in the morning.

Resolves: #2